fix(web): sanitize Gemini JSON ref tool results
