+++ apps/web/src/lib/ai-gateway/experiments/build-direct-provider.ts
@@ -1,10 +1,7 @@
 import { addCacheBreakpoints } from '@/lib/ai-gateway/providers/openrouter/request-helpers';
 import type { CustomLlmApiConfig } from '@kilocode/db';
-import {
-  type GatewayChatApiKind,
-  type Provider,
-  type TransformRequestContext,
-} from '@/lib/ai-gateway/providers/types';
+import { type GatewayChatApiKind, type Provider } from '@/lib/ai-gateway/providers/types';
+import { sanitizeJsonRefToolResults } from '@/lib/ai-gateway/providers/sanitize-json-ref-tool-results';
 
 /**
  * Plain in-memory shape: a `CustomLlmApiConfig` merged with the decrypted
@@ -16,65 +13,6 @@ import {
  * plaintext NEVER touches Postgres, Redis, or any tRPC response.
  */
 export type ResolvedExperimentUpstream = CustomLlmApiConfig & { api_key: string };
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === 'object' && value !== null && !Array.isArray(value);
-}
-
-function renameJsonRefProperties(value: unknown): boolean {
-  if (Array.isArray(value)) {
-    return value.reduce<boolean>(
-      (changed, item) => renameJsonRefProperties(item) || changed,
-      false
-    );
-  }
-
-  if (!isRecord(value)) {
-    return false;
-  }
-
-  let changed = false;
-  for (const [key, nestedValue] of Object.entries(value)) {
-    changed = renameJsonRefProperties(nestedValue) || changed;
-    if (key === '$ref') {
-      delete value.$ref;
-      value._ref = nestedValue;
-      changed = true;
-    }
-  }
-  return changed;
-}
-
-function sanitizeJsonRefContent(content: string): string {
-  try {
-    const result: unknown = JSON.parse(content);
-    return renameJsonRefProperties(result) ? JSON.stringify(result) : content;
-  } catch {
-    return content;
-  }
-}
-
-function sanitizeJsonRefToolResults(context: TransformRequestContext) {
-  if (context.request.kind !== 'chat_completions') {
-    return;
-  }
-
-  for (const message of context.request.body.messages) {
-    if (message.role !== 'tool') {
-      continue;
-    }
-
-    if (typeof message.content === 'string') {
-      message.content = sanitizeJsonRefContent(message.content);
-    } else {
-      for (const part of message.content) {
-        if (part.type === 'text') {
-          part.text = sanitizeJsonRefContent(part.text);
-        }
-      }
-    }
-  }
-}
 
 /**
  * Builds a `Provider` that points directly at a partner-issued upstream.
@@ -117,7 +55,7 @@ export function buildDirectProvider(
         addCacheBreakpoints(context.request);
       }
       if (upstream.sanitize_ref_fields) {
-        sanitizeJsonRefToolResults(context);
+        sanitizeJsonRefToolResults(context.request);
       }
     },
   };
